Move ADNS code into own directory
