+++ keyboards/oddball/rules.mk
@@ -36,4 +36,4 @@ SPLIT_KEYBOARD = yes
 
 # Add trackball support
 POINTING_DEVICE_ENABLE = yes
-SRC += adns.c
+SRC += adns/adns.c
